+++ MPChartLib/com/github/mikephil/charting/listener/BarLineChartTouchListener.java
@@ -589,7 +589,7 @@ public class BarLineChartTouchListener<T extends BarLineChartBase<? extends BarL
         mDecelerationVelocity.x *= mChart.getDragDecelerationFrictionCoef();
         mDecelerationVelocity.y *= mChart.getDragDecelerationFrictionCoef();
 
-        final float timeInterval = (float)(currentTime - mDecelerationLastTime) / 1000.f;
+        final float timeInterval = (currentTime - mDecelerationLastTime) / 1000.f;
 
         float distanceX = mDecelerationVelocity.x * timeInterval;
         float distanceY = mDecelerationVelocity.y * timeInterval;
